factor out serialization pkg
