+++ agenthub/monologue_agent/utils/prompts.py
@@ -5,11 +5,11 @@ from opendevin.core.config import config
 from opendevin.core.exceptions import LLMOutputError
 from opendevin.events.action import (
     Action,
-    action_from_dict,
 )
 from opendevin.events.observation import (
     CmdOutputObservation,
 )
+from opendevin.events.serialization.action import action_from_dict
 
 from . import json
