+++ opendevin/events/action/__init__.py
@@ -1,5 +1,3 @@
-from opendevin.core.exceptions import AgentMalformedActionError
-
 from .action import Action
 from .agent import (
     AgentDelegateAction,
@@ -15,49 +13,6 @@ from .empty import NullAction
 from .files import FileReadAction, FileWriteAction
 from .message import MessageAction
 from .tasks import AddTaskAction, ModifyTaskAction
-
-actions = (
-    CmdKillAction,
-    CmdRunAction,
-    IPythonRunCellAction,
-    BrowseURLAction,
-    FileReadAction,
-    FileWriteAction,
-    AgentRecallAction,
-    AgentFinishAction,
-    AgentRejectAction,
-    AgentDelegateAction,
-    AddTaskAction,
-    ModifyTaskAction,
-    ChangeAgentStateAction,
-    MessageAction,
-)
-
-ACTION_TYPE_TO_CLASS = {action_class.action: action_class for action_class in actions}  # type: ignore[attr-defined]
-
-
-def action_from_dict(action: dict) -> Action:
-    if not isinstance(action, dict):
-        raise AgentMalformedActionError('action must be a dictionary')
-    action = action.copy()
-    if 'action' not in action:
-        raise AgentMalformedActionError(f"'action' key is not found in {action=}")
-    if not isinstance(action['action'], str):
-        raise AgentMalformedActionError(
-            f"'{action['action']=}' is not defined. Available actions: {ACTION_TYPE_TO_CLASS.keys()}"
-        )
-    action_class = ACTION_TYPE_TO_CLASS.get(action['action'])
-    if action_class is None:
-        raise AgentMalformedActionError(
-            f"'{action['action']=}' is not defined. Available actions: {ACTION_TYPE_TO_CLASS.keys()}"
-        )
-    args = action.get('args', {})
-    try:
-        decoded_action = action_class(**args)
-    except TypeError:
-        raise AgentMalformedActionError(f'action={action} has the wrong arguments')
-    return decoded_action
-
 
 __all__ = [
     'Action',
