+++ opendevin/runtime/runtime.py
@@ -2,7 +2,6 @@ from abc import abstractmethod
 
 from opendevin.core.config import config
 from opendevin.events.action import (
-    ACTION_TYPE_TO_CLASS,
     Action,
     AgentRecallAction,
     BrowseURLAction,
@@ -18,6 +17,7 @@ from opendevin.events.observation import (
     NullObservation,
     Observation,
 )
+from opendevin.events.serialization.action import ACTION_TYPE_TO_CLASS
 from opendevin.runtime import (
     DockerExecBox,
     DockerSSHBox,
